requirements.txt
alpaca-py>=0.20.0
pandas>=2.0.0
numpy>=1.26.0
ta>=0.11.0
python-dotenv>=1.0.0
pytz>=2024.1
websockets>=12.0
aiohttp>=3.9.0
pyarrow>=14.0.0
